refactor: nest 아키텍처에 맞게 변경
