+++ src/app.controller.ts
@@ -1,22 +1,11 @@
-import { Controller, Get, HttpStatus, Param, Res } from "@nestjs/common";
+import { Controller, Get, HttpStatus, Res } from "@nestjs/common";
 import { Response } from "express";
+
 import { AppService } from "./app.service";
-import { version } from "../package.json";
-import { IGlasses, ITodos } from "./common/common.interface";
-import { IServiceEndpoint } from "./common/common.endpoint.interface";
+
 @Controller()
-export class AppController implements IServiceEndpoint {
+export class AppController {
   constructor(private readonly appService: AppService) {}
-
-  @Get(":userId/today-todos")
-  getTodayTodos(@Param("userId") userId: string): Promise<ITodos> {
-    return this.appService.getTodayTodos(userId);
-  }
-
-  @Get(":userId/glass")
-  getPlantingGlass(@Param("userId") userId: string): Promise<IGlasses> {
-    return this.appService.getPlantingGlass(userId);
-  }
 
   @Get("ping")
   sendStatusOK(@Res() res: Response): void {
@@ -25,6 +14,11 @@ export class AppController implements IServiceEndpoint {
 
   @Get("version")
   getVersion() {
-    return version;
+    return this.appService.getVersion();
+  }
+
+  @Get("env")
+  getEnv() {
+    return this.appService.getEnv();
   }
 }
