+++ src/todo-service/todo-service.http
@@ -1,7 +1,8 @@
 @Host = http://localhost:3000
+@Router = todo-service
 
 ### 오늘 하루 todo list 조회
-GET {{Host}}/1/today-todos
+GET {{Host}}/{{Router}}/1/today-todos
 
 ### 잔디밭 조회
-GET {{Host}}/2/glass
+GET {{Host}}/{{Router}}/2/glass
